requirements.txt
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-cloud-appengine-logging==1.3.0
google-cloud-audit-log==0.2.5
google-cloud-scheduler==2.11.0
google-crc32c==1.5.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.12.6
grpcio-status==1.54.2
httplib2==0.22.0
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydata-google-auth==1.8.0
pyparsing==3.1.1
python-dotenv==1.0.0
requests==2.31.0
uritemplate==4.1.1
supabase==2.3.1
supafunc==0.3.3
Werkzeug==3.0.1 
cloudevents==1.10.1 
functions-framework==3.5.0
watchdog==3.0.0
huggingface-hub==0.20.3
safetensors==0.4.2
sentence-transformers==2.3.1
tokenizers==0.15.1
transformers==4.37.2
